+++ wikimedia/utilities/exceptions.py
